Move SSO and subscription adapters
