+++ src/adapters/sso/index.ts
@@ -1,3 +1,3 @@
 export * from '@/core/sso/ISsoDataProvider';
 export * from './factory';
-export * from './supabase-adapter';
+export * from './supabase/supabase-adapter';
